requirements.txt
setuptools~=50.3.2
google-cloud-translate~=3.0.1
flask~=1.1.2
flask_restful~=0.3.8
uWSGI~=2.0.19.1
wheel~=0.35.1
flask-limiter~=1.4
pytest~=6.1.2
requests-mock~=1.8.0
pip-tools~=5.4.0
